requirements.txt
streamlit
python-dotenv==0.10.1
requests
google-generativeai